Moved model-updater from vis/ to windshaft-integration/ Also, introduced some adapters to convert rule data to attributes for the legend models
