+++ src/vis/vis.js
@@ -9,7 +9,7 @@ var WindshaftNamedMap = require('../windshaft/named-map');
 var WindshaftAnonymousMap = require('../windshaft/anonymous-map');
 var AnalysisFactory = require('../analysis/analysis-factory');
 var CartoDBLayerGroup = require('../geo/cartodb-layer-group');
-var ModelUpdater = require('./model-updater');
+var ModelUpdater = require('../windshaft-integration/model-updater');
 var LayersCollection = require('../geo/map/layers');
 var AnalysisPoller = require('../analysis/analysis-poller');
 var LayersFactory = require('./layers-factory');
