+++ src/windshaft-integration/legends/rule-to-bubble-legend-adapter.js
@@ -1,0 +1,26 @@
+module.exports = {
+  canAdapt: function (rule) {
+    return rule.prop === 'marker-width';
+  },
+
+  adapt: function (rule) {
+    var values = [
+      rule.stats.min_val
+    ].concat(rule.filters)
+    .concat(rule.stats.max_val);
+
+    var sizes = rule.values;
+    if (rule.mapping === '>') {
+      sizes.unshift(rule['default-value']);
+    } else {
+      sizes.push(rule['default-value']);
+    }
+
+    return {
+      values: values,
+      sizes: sizes,
+      avg: rule.stats.avg_val,
+      state: 'success'
+    };
+  }
+};
