+++ test/unit/windshaft-integration/model-updater.spec.js
@@ -2,7 +2,7 @@ var CartoDBLayer = require('../../../src/geo/map/cartodb-layer');
 var TorqueLayer = require('../../../src/geo/map/torque-layer');
 var LayersCollection = require('../../../src/geo/map/layers');
 var Backbone = require('backbone');
-var ModelUpdater = require('../../../src/vis/model-updater');
+var ModelUpdater = require('../../../src/windshaft-integration/model-updater');
 var WindshaftError = require('../../../src/windshaft/error');
 var CartoDBLayerGroup = require('../../../src/geo/cartodb-layer-group');
 var WindshaftMap = require('../../../src/windshaft/map-base.js');
